# progress bars for EnsembleProblem