@pepijndevos 2023-09-29
Yea that's a fair criticism. I did it like this for a few reasons:

- The alternative would be some complicated thing with channels
- For small ensembles the current default is reasonable, and gives more insight than a single progress bar
- It's fairly easy to aggregate the logs if desired, as the snippet above demonstrates

Maybe it'd be a good idea to contribute this log aggregation somewhere.
We could even have that as the default where in `__solve` we just do `with_logger(sum_logger(current_logger()))`