@pepijndevos 2023-10-02
I've added an internal aggregation step that is enabled by default. This makes sure the current UX isn't horrible, and then we can take it from there.