@pepijndevos 2023-10-04
It turns out that if you have many cores and you set `progress_steps` too low it can cause quite a lot of lock contention, so I made the aggregator use `trylock` to mitigate that.