@pepijndevos 2023-10-06
FTR we're still investigating how to reduce overhead from progress logging, so some changes are still expected.