@pepijndevos 2023-10-06
Alright, performance overhead is now negligible using several strategies:
- use `trylock` to avoid contention on non-critical updates
- use `time` to limit updates to ten per second to limit downstream logging overhead
- use math to avoid summing 20k items
- teach Sundial to respect `progress_steps`

Profile:
![image](https://github.com/SciML/SciMLBase.jl/assets/168609/3ff7eecc-75ad-41de-a325-c8c58e9296e6)

I've also added the requested weakdep on https://github.com/SciML/DiffEqBase.jl/releases/tag/v6.132.0