@ChrisRackauckas 2023-10-28
Even easier, just set ignore derivatives needs to be set on the logging calls so it ignores everything in the function.

Have you done regression testing to check the performance?