@ChrisRackauckas 2023-10-28
https://github.com/SciML/SciMLBase.jl/actions/runs/6532475287/job/17735719200?pr=514 is pointing out that the extra kwargs are causing issues. In order to for this to get in, every downstream solver package needs to support the new keyword arguments. This includes:

- StochasticDiffEq
- JumpProcesses
- DelayDiffEq
- ODEInterfaceDiffEq
- NonlinearSolve
- SimpleNonlinearSolve
- Optimization
- Integrals

It would be fine to set it so that if the value is not the default that you just warn that progress isn't implemented (behind a `if verbose)`) and make sure the kwarg is allowed.