@pepijndevos 2023-10-28
I've never worked with zygote, are there some docs about how to do the derivative or is it a one line thing that would take you less time than finding the docs?

We've tested performance in Cedar and the overhead from logging is minor. I go to quite some length to avoid blocking on a lock or sending more updates than necessary. Julia logging is also such that the message isn't even computed if logging is disabled, so there should be no impact unless you enable progress logging.