@ChrisRackauckas 2023-10-28
Can you show it on something simple like the Lorenz equation with SimpleTsit5?