@pepijndevos 2023-10-30
I'm looking at the docs and all I can find is how to define a rrule? https://fluxml.ai/Zygote.jl/latest/limitations/#Solutions-1
Feels like this should go in ChainRules then?
Or do you mean something else by
>  just set ignore derivatives needs to be set on the logging calls so it ignores everything in the function.