@pepijndevos 2023-11-02
Keeping track of all the PRs:


- [ ]     StochasticDiffEq https://github.com/SciML/StochasticDiffEq.jl/pull/552
- [ ]     JumpProcesses
- [ ]     DelayDiffEq
- [ ]     ODEInterfaceDiffEq
- [ ]     NonlinearSolve
- [ ]     SimpleNonlinearSolve
- [ ]     Optimization
- [ ]     Integrals