@pepijndevos 2023-11-06
Am I going crazy?
https://github.com/SciML/SciMLBase.jl/actions/runs/6769050135/job/18394689941#step:6:579
https://github.com/pepijndevos/SciMLBase.jl/blob/pv/progress/src/SciMLBase.jl#L25