@oxinabox 2023-11-07
ok new chainrules release is out, so restarting CI and it should pass