@oscardssmith 2023-11-07
this pr needs to bump the chainrules version requirement