@ChrisRackauckas 2023-11-07
I don't think this PR ever had a version requirement on ChainRules/ChainRulesCore?