@pepijndevos 2023-11-07
No I added one for DiffEqBase because I thought you wanted that to guarantee `progress_id` is supported but then you removed it again so I'm fine adding a version constraint if it's clear that that's actually what you want