@ChrisRackauckas 2023-11-09
# Before

#=
BenchmarkTools.Trial: 198 samples with 1 evaluation.
 Range (min … max):  22.291 ms … 43.583 ms  ┊ GC (min … max): 0.00% … 43.89%
 Time  (median):     23.324 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   25.246 ms ±  5.548 ms  ┊ GC (mean ± σ):  6.91% ± 12.90%

  ▆█▇▇▅▃▂                                                      
  ███████▆▁▆▅▅▅▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▁▁▅▁▁▁▁▁▁▁▁▅▁▁▅▆▆▆▆▁▇█ ▅
  22.3 ms      Histogram: log(frequency) by time      42.7 ms <

 Memory estimate: 24.42 MiB, allocs estimate: 320057.

BenchmarkTools.Trial: 89 samples with 1 evaluation.
 Range (min … max):  41.775 ms … 83.893 ms  ┊ GC (min … max):  0.00% … 34.04%
 Time  (median):     44.861 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   56.267 ms ± 15.007 ms  ┊ GC (mean ± σ):  22.08% ± 19.77%

   █▄                                                          
  ████▇▄▃▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▄▅▃▄▅▆▇▆▄▄▅▁▁▁▁▃▃ ▁
  41.8 ms         Histogram: frequency by time        80.1 ms <

 Memory estimate: 152.59 MiB, allocs estimate: 1190057.
=#

# After

#=
BenchmarkTools.Trial: 221 samples with 1 evaluation.
 Range (min … max):  19.745 ms … 51.209 ms  ┊ GC (min … max): 0.00% … 56.97%
 Time  (median):     20.502 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.704 ms ±  7.274 ms  ┊ GC (mean ± σ):  8.31% ± 14.29%

  ▇█▅▂▁                                                        
  █████▄▅▁▁▁▄▁▄▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▄▇▄▅▁▆ ▅
  19.7 ms      Histogram: log(frequency) by time      50.6 ms <

 Memory estimate: 24.42 MiB, allocs estimate: 320076.

BenchmarkTools.Trial: 31 samples with 1 evaluation.
 Range (min … max):  141.556 ms … 179.966 ms  ┊ GC (min … max):  4.08% … 11.54%
 Time  (median):     166.417 ms               ┊ GC (median):    11.92%
 Time  (mean ± σ):   166.345 ms ±   8.579 ms  ┊ GC (mean ± σ):  12.01% ±  2.93%

                           ▃         █  ▃  ▃                  ▃  
  ▇▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇█▁▁▁▁▇▇▁▇▇█▁▇█▇▇█▁▇▇▇▇▇▁▁▁▇▇▁▇▁▇▇▁▁█ ▁
  142 ms           Histogram: frequency by time          180 ms <

 Memory estimate: 220.28 MiB, allocs estimate: 1785509.
=#